Changes to Java code types and moving to sum for performance times

+++ java/primes.java
@@ -1,14 +1,4 @@
-/**
- * Ganesh Bhat
- * https://github.com/ganeshkbhat/fastprimecalculations
- * Fastest method of Prime Number Calculations
- * Functions and Alternate ways for calculation of Prime Numbers
- * 
- * ** prime(n) ** is the Recommended Way for calculation checks of prime numbers
- * 
-*/
 
-import java.lang.*;
 /**
  * Ganesh Bhat
  * https://github.com/ganeshkbhat/fastprimecalculations
@@ -22,8 +12,9 @@ import java.lang.*;
 import java.lang.*;
 
 public class Primes {
-    public static boolean primes(int n)  {
-        int count = 0;
+    
+    public static boolean primes(long n)  {
+        long count = 0;
         if ( (n == 2 || n == 3 || n == 5 || n == 7) ) {
             System.out.println("count: Prime Unconventional way for " + n + " is " + count + "\n"); 
             return true;
@@ -34,13 +25,13 @@ public class Primes {
         }
         double factorminusone = (n-1)/6;
         double factorplusone = (n+1)/6;
-        if ( !( factorminusone > (int)factorminusone ) || !( factorplusone > (int)factorplusone ) ){
-            for (int i=1; i <= n; i++) {
+        if ( !( factorminusone > (long)factorminusone ) || !( factorplusone > (long)factorplusone ) ){
+            for (long i=1; i <= n; i++) {
                 count += 1;
-                int factorsix = (i * 6);
+                long factorsix = (i * 6);
                 double fivebase = n / (5 + factorsix);
                 double sevenbase = n / (7 + factorsix);
-                if (((fivebase > 1) && !( fivebase > (int) fivebase ) ) || ((sevenbase > 1) && !( sevenbase > (int) sevenbase ) )){
+                if (((fivebase > 1) && !( fivebase > (long) fivebase ) ) || ((sevenbase > 1) && !( sevenbase > (long) sevenbase ) )){
                     System.out.println("count: Prime Unconventional way for " + n + " is " + count + "\n"); 
                     return false;
                 }
@@ -56,11 +47,14 @@ public class Primes {
     }
 
     public static void main(String args[]) {
+
         System.out.println("Importing Primes Recommended Way ");
+
         // System.out.println("Is Prime Unconventional way for 83: prime is  " + " is " + primes(83));
         // System.out.println("Is Prime Unconventional way for 169: prime is  " + " is " + primes(169));
         // System.out.println("Is Prime Unconventional way for 167: prime is  " + " is " + primes(167));
         // System.out.println("Is Prime Unconventional way for 300530164787: prime is " +  prime(300530164787));
+
     }
 }
